betty fixes

+++ 0x1D-search_algorithms/search_algos.h
@@ -4,7 +4,8 @@
 #include <stdio.h>
 
 int jump_search(int *array, size_t size, int value);
-int _search(int *array, short jump, int size, int hi, int low, int value);
+int _search(int *array, size_t jump, size_t size, size_t hi, size_t low,
+	    int value);
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
 int _split(int *array, size_t hi, size_t low, int value);
